requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
django-environ==0.9.0